added PID, IIR, time sync functions

+++ stspin/steval-spin3204-6step/6step.c
@@ -788,7 +788,7 @@ void pwm6step_set_duty_cycle(int32_t duty_cycle) {
 }
 
 void pwm6step_set_duty_cycle_f(float duty_cycle_pct) {
-    pwm6step_set_duty_cycle((uint16_t) (duty_cycle_pct * (float) UINT16_MAX));
+    pwm6step_set_duty_cycle((int32_t) (duty_cycle_pct * (float) UINT16_MAX));
 }
 
 void pwm6step_brake(uint16_t braking_force) {
